updates for linux

+++ main.py
@@ -1,11 +1,14 @@
 
+import glob
+import sys
+import subprocess
 import os
 import pyperclip
 
 from tkinter import *
 from tkinter import filedialog      # for browse window (adding path)
 
-from PIL import Image               # PILLOW import has to be after the tkinter impoert (Image.open will not work: 'Image has no attributesm open')
+from PIL import Image               # PILLOW import has to be after the tkinter import (Image.open will not work: 'Image has no attribute open')
 from PIL import ImageTk
 
 from pathlib import Path
@@ -23,6 +26,9 @@ field_background_color = settings_data['field_background_color']
 font_style = settings_data['font_style']
 font_size = settings_data['font_size']
 font_color = settings_data['font_color']
+working_directory = os.path.dirname(__file__)
+os_linux: bool = sys.platform == 'linux'
+
 
 
 # WINDOW
@@ -36,9 +42,9 @@ window.geometry(f'{window_width}x{window_length}+%d+%d' % (screen_width/2-275, s
 window.resizable(0,0)   # locks the main window
 window.configure(background=settings_data['background_color'])
 # ICON
-working_directory = os.path.dirname(__file__)
-path_icon = Path(working_directory, "skin", "icon.ico") 
-window.iconbitmap(path_icon)
+if not os_linux:
+    path_icon = Path(working_directory, "skin", "icon.ico") 
+    window.iconbitmap(path_icon)
 # RECTANGLE
 canvas_color = settings_data['background_color']
 canvas_frame_color = settings_data['canvas_frame_color']
@@ -47,7 +53,10 @@ canvas.create_rectangle(5-1, 5+2, window_width-5, window_length-5, outline=canva
 canvas.pack()
 # BUTTON SIZE
 button_height = 1
-button_width = 10
+if os_linux:
+    button_width = 7
+else:
+    button_width = 10
 # SEARCH FIELD LENGTH
 search_field_length = 40
 
@@ -80,6 +89,15 @@ class Fields:
                     foreground=font_color, 
                     background=self.background, 
                     font=(font_style, font_size))
+
+
+def get_path_yt_dlp():
+    if os_linux:
+        path_yt_dlp = 'yt-dlp'
+    else:    
+        settings_data = settings.open_settings()
+        path_yt_dlp = settings_data['path_yt_dlp']
+    return path_yt_dlp
 
 
 
@@ -129,9 +147,10 @@ def button_get_url_actions():
             settings_data['video_duration'] = ""
             settings.save_settings(settings_data)
             path_thumbnail = Path(working_directory, "thumbnail", "thumbnail.png")
-            os.remove(path_thumbnail)
+            if path_thumbnail.is_file():
+                os.remove(path_thumbnail)
         except:
-            pass
+            print('Remove previous information failed')
 
     def get_url():
         settings_data['video_url'] = pyperclip.paste()
@@ -139,7 +158,7 @@ def button_get_url_actions():
 
     # GET INFORMATION > INFO.TXT
     def save_info():
-        path_yt_dlp = settings_data['path_yt_dlp']
+        path_yt_dlp = get_path_yt_dlp()
         link = settings_data['video_url']
         info_path = Path(working_directory, "info", "info.txt") 
         parameter = f'--print id --get-title --get-duration --restrict-filenames --quiet'
@@ -167,7 +186,7 @@ def button_get_url_actions():
     def save_thumbnail():
         if settings_data['video_title'] != "":
             # YT-DLP
-            path_yt_dlp = settings_data['path_yt_dlp']
+            path_yt_dlp = get_path_yt_dlp()
             link = settings_data['video_url']
             path = Path(working_directory, "thumbnail")
             parameter = f'--skip-download -o %(NAME)s --write-thumbnail --convert-thumbnails png --paths "{path}" --quiet' % {'NAME': "thumbnail"}
@@ -221,19 +240,23 @@ def button_get_url_actions():
             text_position(video_title_field)
 
     def yt_dlp_path_valuation():
-        if settings_data['path_yt_dlp'] != "" and "mandatory" not in settings_data['path_yt_dlp']:
-            pass
+        if os_linux:
+            return True
         else:
-            messages.error_pop_up('Error','no_yt_dlp')
+            if os.path.isfile(settings_data['path_yt_dlp']):
+                return True
+            else:
+                messages.error_pop_up('Error','no_yt_dlp')
+                return False
 
-    yt_dlp_path_valuation(),
-    remove_pre_info(),
-    get_url(),
-    save_info(),
-    extract_info(),
-    save_thumbnail(),
-    display_thumbnail(),
-    display_info()     
+    if yt_dlp_path_valuation():
+        remove_pre_info(),
+        get_url(),
+        save_info(),
+        extract_info(),
+        save_thumbnail(),
+        display_thumbnail(),
+        display_info()     
 
 button_get_url_instance = Buttons("Get URL", lambda: [button_get_url_actions()])
 button_get_url = button_get_url_instance.create()
@@ -257,8 +280,16 @@ for item in av_options:
 av_options_roll_down_clicked = StringVar()
 av_options_roll_down_clicked.set("Save as")    
 av_options_roll_down = OptionMenu( window, av_options_roll_down_clicked, *av_options_list, command=None)     
-av_options_roll_down.configure(foreground=font_color, background=background_color, activeforeground = font_color, activebackground=background_color, highlightbackground=background_color)
-av_options_roll_down['menu'].configure(foreground=font_color, background=background_color, activebackground='grey')
+av_options_roll_down.configure(
+    foreground=font_color,
+    background=background_color,
+    activeforeground = font_color,
+    activebackground=background_color,
+    highlightbackground=background_color)
+av_options_roll_down['menu'].configure(
+    foreground=font_color,
+    background=background_color,
+    activebackground='grey')
 
 
 ## START - BUTTON
@@ -266,7 +297,7 @@ def start():
 
     ## VALUE CHECKS - MESSAGES
     # DESTINATION FOLDER ADDED
-    if destination_field.get("1.0", "end-1c") == "":
+    if not os.path.isdir(destination_field.get("1.0", "end-1c")):
         messages.error_pop_up('Error','destination_folder')
         return
     
@@ -282,7 +313,7 @@ def start():
     
     # YT-DLP    - already checked with the "Get URL", why checking here too: unique scenario, YT-DLP added, rest of the req.es added(destination, URL) and YT-DLP removed before click START
     settings_data = settings.open_settings()    # otherwise it will use the value pulled from DB at the start of the program
-    if settings_data['path_yt_dlp'] == "":
+    if not os_linux and settings_data['path_yt_dlp'] == "":
         messages.error_pop_up('Error', 'no_yt_dlp')
         return
     
@@ -298,7 +329,7 @@ def start():
     link = settings_data['video_url']
     selected_resolution = av_options[av_selected]       #av_options['720p']
     path = destination_field.get("1.0", "end-1c")
-    path_yt_dlp = settings_data['path_yt_dlp']
+    path_yt_dlp = get_path_yt_dlp()
     
     # FFMPEG PATH 
     if settings_data['path_ffmpeg'] != "":          # if ffmpeg not added to the windows path - ffmpeg path browse field is used
@@ -319,10 +350,15 @@ def start():
 
     # UPDATE THE FILE`S MODIFIED DATE ATTRIBUTE TO THE CURRENT DATE
     def change_modified_date():
-        os.chdir(path)
         video_ID = settings_data['video_ID']
-        update_date = f'copy /b *"{video_ID}"* +,,'
-        os.system(update_date)
+        if os_linux:
+            file_path = glob.glob(f'{path}/*{video_ID}*')[0]
+            subprocess.Popen(f'touch "{file_path}"', shell=True)
+        else:
+            os.chdir(path)
+            update_date = f'copy /b *"{video_ID}"* +,,'
+            os.system(update_date)
+    
     change_modified_date()
 
 button_start_instance = Buttons("START", lambda: [start()])
